added conformer model back and refactored to be better shared
