+++ config/model/conformer_ctc.yaml
@@ -1,0 +1,19 @@
+# @package _global_
+module:
+  _target_: emg2qwerty.conformer_lightning.ConformerCTCModule
+  in_features: 96  # freq * channels = (n_fft // 2 + 1) * 16
+  mlp_features: [256]
+  d_model: 128
+  nhead: 4
+  num_layers: 4
+  ff_expansion_factor: 2
+  conv_kernel_size: 15
+  conv_expansion_factor: 1
+  dropout: 0.2
+  ctc_weight: 0.7
+  ce_weight: 0.3
+
+datamodule:
+  _target_: emg2qwerty.lightning.WindowedEMGDataModule
+  window_length: 8000  # 4 sec windows for 2kHz EMG
+  padding: [1800, 200]  # 900ms past context, 100ms future context
