+++ config/model/conformer_decoder.yaml
@@ -1,18 +1,18 @@
 # @package _global_
 module:
   _target_: emg2qwerty.conformer_lightning.ConformerDecoder
-  in_features: 96  # freq * channels = (n_fft // 2 + 1) * 16
+  in_features: 96  # freq * n_mels
   mlp_features: [256]
-  d_model: 256
-  nhead: 16
+  d_model: 128
+  nhead: 4
   num_encoder_layers: 8
-  num_decoder_layers: 8
+  num_decoder_layers: 4
   ff_expansion_factor: 2
   conv_kernel_size: 7
   conv_expansion_factor: 1
   dropout: 0.1
 
 datamodule:
-  _target_: emg2qwerty.lightning.WindowedEMGDataModule
-  window_length: 8000  # 4 sec windows for 2kHz EMG
+  _target_: emg2qwerty.conformer_lightning.WindowedEMGDataModule
+  window_length: 2000  # 4 sec windows for 2kHz EMG
   padding: [0, 0]  # No padding, [1800, 200]  # 900ms past context, 100ms future context
