--- emg2qwerty/custom_modules.py
@@ -1,7 +1,0 @@
-from collections.abc import Sequence
-
-import torch
-from torch import nn
-
-
-# TODO: build modules for our custom model
